Suppress expected warnings, more threading for GUI
Main warnings from fermi_libraries.common_functions.rebinning are now
suppressed (we run them on length-1 arrays when initializing the GUI).

Some parts of the GUI which prepares the data for graphing are now put
in threads. They take a small but noticeable time, so the GUI froze a
bit (~0.5s). It still freeze, so more processing blocks must be put into
threads as well (e.g. for the TOF part).

Added more parts to the GUI including the fitted and inverted Abel
images, and some other things.