Revert to Run() object for TOF data
There is a problem running this GUI on Linux, where the VMI portion
runs, but the TOF portion gets stuck when attempting to open/read the
HDF5 files. This only happens when running the get_new_tof_data() with
two simultaneous conditions:
    - running the get_new_tof_data() in a Worker
    - using the MultithreadRun class for data compilation.

If either of these is False, then it works. I suspect it has something
to do with QThreadpool, so I intend to use separate QThreads for the VMI
and TOF in a future update.