Implemented caching info for Run methods
- New functions with cache info
- New test for cache saving/loading

Many methods for Run and RunSet were rewritten with giving the cache
information in mind.

Include a new test for cache saving and loading, which directly looks at the
files that were either saved or loaded in test_caching_function.py. This is a
more specific test than the previous false positive test in basic_caching.py.