Extenal verification of cache saves
- Cache test with more checks
- Fix for common.functions.find_subdir

Cache test has extra checks where the the of file modification is
checked (in case the cache files already exist, but are somehow not
overwritten).

There was non-deterministic behaviour when using
common_functions.find_subdir while running the cache tests. This is now
fixed in general, and works properly with standard formatting.